requirements.txt
google-generativeai
mysql-connector-python
scikit-learn
pandas
python-dotenv